Crate lazy_vec

A "lazy vector" is a self-initializing vector: it can be created in constant time, but still has constant-time read and write. It initializes an element on first write, and occupies space proportional to the number of written elements.

Examples

use lazy_vec::LazyVec;
 
let mut a: LazyVec<i8> = LazyVec::new();
a[77] = -12i8;
assert_eq!(a[77], -12i8);

Structs

LazyVec
This opaque structure stores a lazy vector.