Struct lazy_vec::LazyVec

pub struct LazyVec<T> { /* fields omitted */ }

This opaque structure stores a lazy vector.

Methods

impl<T: Copy> LazyVec<T>

pub fn new() -> LazyVec<T>

Allocate a new empty LazyVec.

pub fn with_capacity(cap: usize) -> LazyVec<T>

Allocate a new empty LazyVec with the given starting index capacity.

pub fn cap(&self) -> usize

Number of elements there is currently notional capacity for, including uninitialized ones.

pub fn len(&self) -> usize

Number of elements notionally stored, including uninitialized ones.

pub fn value_ref(&self, i: usize) -> &T

Return a reference to the value at the given index.

XXX For now, panic on attempt to read from an uninitialized element.

pub unsafe fn value_ref_mut(&mut self, i: usize) -> &mut T

Return a mutable reference to the value at index i. If no value previously existed, this will return a reference to uninitialized memory, making it unsafe.